64-bit fixes (the parts of lib/ in use at work are now fully operational in 64-bit mode) - x86_x64.cpp: fix: test ARCH_AMD64 instead of ARCH_IA32 because both are defined on x64 - manifest.cpp: fix processor architecture - wdbg_heap.cpp: only enable on ia32 - winit, wstartup: take (lack of) name mangling into account
compiler.h: add STRINGIZE and DECORATED_NAME for use by winit
dir_watch: fix: smart pointers were never actually released due to a remaining reference

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6375 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/os/win/wdll_ver.cpp
@@ -52,7 +52,7 @@ static LibError ReadVersionString(const OsPath& modulePathname_, char* out_ver, 
 		WARN_RETURN(ERR::FAIL);
 
 	u16* lang;	// -> 16 bit language ID, 16 bit codepage
-	size_t lang_len;
+	UINT lang_len;
 	const BOOL ok = VerQueryValue(mem.get(), "\\VarFileInfo\\Translation", (void**)&lang, &lang_len);
 	if(!ok || !lang || lang_len != 4)
 		WARN_RETURN(ERR::FAIL);
@@ -60,7 +60,7 @@ static LibError ReadVersionString(const OsPath& modulePathname_, char* out_ver, 
 	char subblock[64];
 	sprintf(subblock, "\\StringFileInfo\\%04X%04X\\FileVersion", lang[0], lang[1]);
 	const char* in_ver;
-	size_t in_ver_len;
+	UINT in_ver_len;
 	if(!VerQueryValue(mem.get(), subblock, (void**)&in_ver, &in_ver_len))
 		WARN_RETURN(ERR::FAIL);
